[GH Idx] Extract GitFileInfo class
